requirements.txt
python-dotenv==1.0.1
docx2txt
instructor
openai